+++ src/cg/assembler/codeextract.cpp
@@ -202,20 +202,23 @@ vector<string> assembler::ce::getlinks(const vector<viua::cg::lex::Token>& token
     return links;
 }
 
+static bool looks_like_name_definition(Token t) {
+    return (t == ".function:" or t == ".closure:" or t == ".block:" or t == ".signature:" or t == ".bsignature:");
+}
 static vector<string> get_instruction_block_names(const vector<Token>& tokens, string directive, void predicate(Token) = [](Token){}) {
     vector<string> names;
+    vector<string> all_names;
     map<string, Token> defined_where;
 
     const auto limit = tokens.size();
     string looking_for = ("." + directive + ":");
     for (decltype(tokens.size()) i = 0; i < limit; ++i) {
-        if (tokens[i].str() == looking_for) {
+        if (looks_like_name_definition(tokens.at(i))) {
             ++i;
             if (i >= limit) {
                 throw tokens[i-1];
             }
 
-            predicate(tokens.at(i));
             if (defined_where.count(tokens.at(i)) > 0) {
                 throw viua::cg::lex::TracedSyntaxError()
                     .append(viua::cg::lex::InvalidSyntax(tokens.at(i), ("duplicated name: " + tokens.at(i).str())))
@@ -223,7 +226,10 @@ static vector<string> get_instruction_block_names(const vector<Token>& tokens, s
                     ;
             }
 
-            names.emplace_back(tokens.at(i).str());
+            if (tokens.at(i-1) == looking_for) {
+                predicate(tokens.at(i));
+                names.emplace_back(tokens.at(i).str());
+            }
             defined_where.emplace(tokens.at(i), tokens.at(i));
         }
     }